requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-environ==0.9.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0
